Patch bulk loader relation bug from http://open.silverstripe.org/ticket/6472
Use Case:
User is in a ModelAdmin? view in the backend and wants to import Data. Before
the import he checks "Clear database before import". The managed class has a
has_one relation to another class with an attribute "ID".

Error:
The BulkLoader? class (located in "/sapphire/dev") crashes with the exception
"Column 'ID' in field list is ambiguous".