BUGFIX: Fixes #7003
Changed the calculation of the iframe size so it would be the correct
height. Altered the styles of the view details area to match design.
Changed the background color of the title on the file header so it
wouldn't look editable (as recommended by Felipe). Then fixed the
iframe to work in IE7 and IE8.