DOC Gave easier instructions for would-be authors
The use of GitHub for documentation isn't ideal in a world full of wikis,
but we do it because it has other substantial advantages. However, we
shouldn't assume that every would-be author is comfortable using pull requests.

This change presents the edit-on-github interaction as the most straightforward
one, suggesting that advanced users may which to check out locally.