API Fixing id attribute on FieldHolder form templates
Previously the id attribute was set to $Name, but this does not
clean up the value to satisfy HTML validation for an id attribute.
Additionally, the holder id would conflict with the field input id.

This fixes the FieldHolder templates to use "{$ID}_holder" for the
id attribute, which is namespaced to the form ID to avoid conflicts,
and any characters not allowed in the attribute are removed.