Fix broken link in docs/en/index.md
The "Introduction to PHP" link was broken, I’ve changed it and points now to "Zend – PHP 101: PHP For the Absolute Beginner".

The "Built-in Page Controls" has not been updated to 3.0. I don’t know how do you want to manage this, so I suggest to follow the 2.4 link in the target page for beginners to not be too lost.