BUG: Force gridfield to wrap on long content
Make sure gridfield cells wrap. Added word-warp, so long words will
break apart to wrap, and a max-width.

The max-width number is semi-arbitrary. In practice it does not act as
a real max-width, but as some sort of limit on the percentage.
Unfortunately, this fix has no effect at all on any version of IE.