BUG Using GLOB for case sensitive matches in SQLite3
As opposed to LIKE, the GLOB operator is case sensitive by default
in SQlite3. It uses "*" instead of "%" for wildcards,
which necessitated a new SearchFilter->getWildcard() method.

SQlite3 doesn't support per-term modifiers,
COLLATE BINARY LIKE is case insensitive by default
unless the field collation is set up accordingly.
There's connection-level modifiers (PRAGMA case_sensitive_like = true),
but that would affect all comparisators in the executed query.

This code should be contained in the DB driver instead,
but needs more refactoring, there's a *lot* of duplication
in there already. For now it fixes this
aspect of the SQlite3 builds, which enables us
to work with a more responsive test suite again.