PHP 5.3.2 recommended configuration
I was checking the requirements, I wanted to know more about the following and I came across the php website where the following functions are deprecated in PHP 5.3. Since we use PHP 5.3, there is no need to have them being recomended in the doc?

Source: php.net

Safe mode:
This feature has been DEPRECATED as of PHP 5.3.0 and REMOVED as of PHP 5.4.0.

magic_quotes_gpc
This feature has been DEPRECATED as of PHP 5.3.0 and REMOVED as of PHP 5.4.0.