Passing tests
We were having a test failure due to the change in the way cachekeys
were generated.

The ResfulServiceTest had the logic for generating the cachepath hard
coded in the test class, so this has been pulled into a function on
RestfulService, allowing it to be changed in one place and remain
consistent.