FIX Allow comments in static var declarations
At the moment, comments inside the declaration of static vars are
causing an error.

eg:
```php
protected static
/**
* The URL to access the service
*/
$endpoint = '',
/**
* Our username to access the service
*/
$username = '',
/**
* Our secret key to access the service
*/
$secret = '';
```
Throws an error. This patch fixes that.

It's realistic that this could happen too:
```php
public static /* comment */ $var = '';
```
Though I haven't tested that with this patch.

There is still an issue with this declaration:
```php
public static $var = <<<VAR
Some value
VAR;
```

I'm not certain this is the best solution, so happy to hear thoughts.