GenericTemplateGlobalProvider::getDataList() shouldn't sort by default
The common use of this method is to fetch a single aggregate value
for use in partial caches within templates. Sorting aggregates in this
scenario will break MSSQL and PostgreSQL as it requires the sort be in
the GROUP BY clause as well, but it's not necessary to do any sorting
in this context anyway.