BUG GridField action onclick assuming URL without "?" character
Sometimes GridState URL will have a query string already appended,
e.g. "?local=en_GB", but the GridField javascript assumes there won't
be.

The result would be an action URL like:
"/my/url?locale=en_US?RelatedPage=123" which is obviously invalid.

The fix is to check that "?" exists in the GridState URL and use
"&" instead to connect the URLs if that's the case.