FIX: Prevent UploadField folder selection from being too tall for the iframe
This does make the dropdown menu fairly short (and can result in a large amount of scrolling if lots of folders are present), but is an improvement on not being able to select them at all.

Most of the asset management should be done in the files section anyway.

Add comment for arbitrary value